Checkout to a named branch in submodule test
The checkout to named branch requires the call to 
deleteBranchIfExists(true) because of the side effect
that has of calling git checkout with the "-f" argument.

There is no branch to delete, since it does not exist at
that time, but without the side effect of the "-f" argument,
the call to "git checkout" fails.