Make JGit submodule checkout behavior match CliGitAPIImpl
The CliGitAPIImpl checkout() method defaults to perform a submodule
init immediately after the checkout. It appears to be added as an
extension to the behavior. It is unclear (to me) why that same
extension is not being applied to the JGitAPIImpl checkout() method.

I suspect that means there is a bug in the JGitAPIImpl because it is
not using extensions in the same way as CliGitAPIImpl.