Assert condition of modules/firewall directory at various tests
The modules/firewall directory is a module in the tests/getSubmodules
branch, and does not exist in the tests/notSubmodules branch. That
makes it different from the modules/ntp directory, which is in the
tests/getSubmodules branch as a module, and is in the
tests/notSubmodules branch as a directory.