Change proxy environment variable logic
Changed the logic that sets the http_proxy and https_proxy environmental
variables. Now if there is already an environmental variable set (like
in the node configuration or at the os layer on build user login)
the git-client-plugin will not set it to be what is found in the
"proxy settings" section of the jenkins configuration. This should
mean that for the most part there would be no change in behaviour.

The only difference this would make is if someone has gone to the trouble
of setting these parameters at the node level (or at the o/s layer by
setting the enviornment variables for the build user). If someone has
gone to the trouble of setting the parameters in either of these two
ways then it's likely (but not guaranteed) that they would had set these
parameters correctly.