[JENKINS-44041] Fix windows git clone failures on certain paths
If the Windows command line git program is passed a GIT_SSH (or
GIT_ASKPASS) value which includes a space character, an open parenthesis,
a closed parenthesis, or a percent character, the command line git
program fails to clone if it is trying to clone a username / password
protected repository over http or https.

A similar problem was seen when using passphrase protected private keys
with ssh instead of username / password with http.

Rather than continuing to try to guess which characters will cause
command line git to misbehave on Windows, adopt the same solution
as was earlier used for space characters in the absolute path.  If a
character is detected which might cause trouble for command line git,
use the system temporary directory instead of the temporary directory
adjacent to the workspace.