Test credentials only in command line git
Some of the providers seem to block requests or throttle them if
multiple requests arrive for the same repository, especially when those
requests are comming from different implementations.

Since this pull request is only changing CliGit, there is little benefit
to testing JGit implementations.

Can't merge this as a final change, but can use it for evaluation.