Skip JMH benchmarks based on commit message content
The Jenkinsfile already requires a commit message that includes the word
"benchmark" or "Benchmark".  This change makes the unit test have the
same condition.  JMH benchmarks will be run in CI if and only if the word
"[Bb]enchmark" is included in one of the last 3 messages.

Write empty JMH report file for archiver.