Support the octopus merge strategy
Added showChangedPaths: Given a Revision, show all files changes
similar to git diff --name-only, so that it can be used to
accurately identify changed paths for include/exclude regions.

JGit does not seem to have an implementation of an octopus merge, so
the test falls back to performing the octopus merge with command line
git.

Tests used to show places where the behavior is different between
command line git and JGit implementation of showChangedPaths, and
where the results are unexpected for both implementations.

The showChangedPaths() method for CliGitAPIImpl returns
duplicates in the list, while the JGit implementation only
returns the matching items