Reduce failure rate of GitClientTest.testSparseCheckoutWithCliGitLFS()
Skips the test if a credential file is not found.  Applies the credential
in all cases if it is found.  Even with the credential, the test still
fails about 20% of the time.  That's better than the 100% failure rate
that I see without the credential.

Needs more investigation after I return from vacation.